requirements.txt
httpx
telethon
tqdm
python-dotenv
psycopg2-binary
